+++ src/rhodes/states/__init__.py
@@ -1,17 +1,21 @@
 """"""
 import json
-from typing import Any, Dict, Optional, Iterable
+from typing import Any, Dict, Iterable, List, Optional, TypeVar
 
 import attr
 from attr.validators import deep_iterable, deep_mapping, instance_of, optional
-from troposphere import GetAtt, Ref, Sub, awslambda, stepfunctions
+from troposphere import Sub
 
 from rhodes._converters import convert_to_json_path
-from rhodes._util import RequiredValue, require_field
-from rhodes._validators import is_valid_arn, is_valid_timestamp
+from rhodes._runtime_types import TASK_RESOURCE_TYPES
+from rhodes._serialization import serialize_name_and_value
+from rhodes._util import RHODES_ATTRIB, RequiredValue, require_field
+from rhodes._validators import is_valid_timestamp
 from rhodes.choice_rules import ChoiceRule
 from rhodes.exceptions import InvalidDefinitionError
-from rhodes.structures import ContextPath, JsonPath
+from rhodes.structures import JsonPath, Parameters
+
+from ._parameters import _catch_retry, _input_output, _next_and_end, _parameters, _result_path, task_type
 
 __all__ = (
     "State",
@@ -28,79 +32,16 @@ __all__ = (
 )
 
 
-def _getatt_arn(value: str) -> str:
-    return f"${{{value}.Arn}}"
-
-
-def _ref(value: str) -> str:
-    return f"${{{value}}}"
-
-
-def _serialize_troposphere_value(value):
-    # Inject appropriate Ref/GetAtt for Troposphere
-    if isinstance(value, awslambda.Function):
-        return _getatt_arn(value.title)
-
-    if isinstance(value, stepfunctions.Activity):
-        return _ref(value.title)
-
-    if isinstance(value, Ref):
-        return _ref(value.data["Ref"])
-
-    if isinstance(value, GetAtt):
-        return _getatt_arn(value.data["Fn::GetAtt"][0])
-
-    return value
-
-
-def _serialize_name_and_value(*, name: str, value: Any) -> [str, Any]:
-    value = _serialize_troposphere_value(value)
-
-    if isinstance(value, (JsonPath, ContextPath)):
-        return name, str(value)
-
-    if hasattr(value, "to_dict") and callable(value.to_dict):
-        return name, value.to_dict()
-
-    return name, value
-
-
 @attr.s(eq=False)
 class State:
     """"""
 
     title: str = attr.ib(validator=instance_of(str))
-    # TODO: These need to be required and exclusive OR
-    Next: str = attr.ib(default=None, validator=optional(instance_of(str)))
-    # TODO: End=False is invalid
-    End: bool = attr.ib(default=None, validator=optional(instance_of(bool)))
-    Comment: Optional[str] = attr.ib(default=None, validator=optional(instance_of(str)))
-    InputPath: Optional[JsonPath] = attr.ib(default=None, validator=optional(instance_of(JsonPath)), converter=convert_to_json_path)
-    OutputPath: Optional[JsonPath] = attr.ib(default=None, validator=optional(instance_of(JsonPath)), converter=convert_to_json_path)
 
-    def then(self: "State", next_state: "State") -> "State":
-        if self.End:
-            raise InvalidDefinitionError(
-                "Cannot set state transition." f"State {self.title!r} already has an end condition."
-            )
-
-        self.member_of.add_state(next_state)
-        # TODO: set reference rather than extracting name
-        self.Next = next_state.title
-        return next_state
-
-    def end(self: "State") -> "State":
-        if self.Next:
-            raise InvalidDefinitionError(
-                "Cannot set end condition." f"State {self.title!r} already has a state transition."
-            )
-
-        self.End = True
-
-        return self
+    Comment: Optional[str] = RHODES_ATTRIB(validator=optional(instance_of(str)))
 
     member_of = None
-    _required_fields = []
+    _required_fields: Iterable[RequiredValue] = []
     __setup_complete = False
 
     def __attrs_post_init__(self):
@@ -144,7 +85,7 @@ class State:
             if value is None:
                 continue
 
-            new_name, new_value = _serialize_name_and_value(name=field.name, value=value)
+            new_name, new_value = serialize_name_and_value(name=field.name, value=value)
 
             self_dict[new_name] = new_value
 
@@ -161,18 +102,17 @@ class StateMachine:
     ]
     __setup_complete = False
 
-    # TODO: Map of States
-    States = attr.ib(
+    States = RHODES_ATTRIB(
         default=attr.Factory(dict),
         validator=deep_mapping(key_validator=instance_of(str), value_validator=instance_of(State)),
     )
     # TODO: Name of State
-    StartAt = attr.ib(default=None, validator=optional(instance_of(str)))
-    Comment = attr.ib(default=None, validator=optional(instance_of(str)))
+    StartAt = RHODES_ATTRIB(validator=optional(instance_of(str)))
+    Comment = RHODES_ATTRIB(validator=optional(instance_of(str)))
     # TODO: MUST be 1.0 if provided
-    Version = attr.ib(default=None, validator=optional(instance_of(str)))
+    Version = RHODES_ATTRIB(validator=optional(instance_of(str)))
     # TODO: MUST be non-negative
-    TimeoutSeconds = attr.ib(default=None, validator=optional(instance_of(int)))
+    TimeoutSeconds = RHODES_ATTRIB(validator=optional(instance_of(int)))
 
     def __attrs_post_init__(self):
         self.__setup_complete = True
@@ -227,64 +167,13 @@ class StateMachine:
         return first_state
 
 
-class Parameters:
-    def __init__(self, **kwargs):
-        self._map = kwargs
-
-    def to_dict(self) -> Dict[str, Any]:
-        def _inner():
-            for name, value in self._map.items():
-                new_name, new_value = _serialize_name_and_value(name=name, value=value)
-
-                if isinstance(value, JsonPath) or isinstance(value, ContextPath):
-                    # If you manually provide path strings, you must manually set the parameter suffix.
-                    if not new_name.endswith(".$"):
-                        new_name += ".$"
-
-                yield new_name, new_value
-
-        return dict(_inner())
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}({', '.join(f'{name}={value!r}' for name, value in self._map.items())})"
-
-
-def _parameters(cls):
-    cls.Parameters = attr.ib(default=None, validator=optional(instance_of(Parameters)))
-
-    return cls
-
-
-def _result_path(cls):
-    cls.ResultPath = attr.ib(default=None, validator=optional(instance_of(JsonPath)), converter=convert_to_json_path)
-
-    return cls
-
-
-def _catch_retry(cls):
-    cls.Retry = attr.ib(default=None)
-    cls.Catch = attr.ib(default=None)
-
-    return cls
-
-
 @attr.s(eq=False)
 @_parameters
 @_result_path
+@_input_output
+@_next_and_end
 class Pass(State):
-    Result = attr.ib(default=None)
-
-
-def task_type(cls):
-    cls = _catch_retry(cls)
-    cls = _result_path(cls)
-
-    # TODO: Timeout MUST be positive
-    cls.TimeoutSeconds = attr.ib(default=None, validator=optional(instance_of(int)))
-    # TODO: HeartbeatSeconds MUST be positive
-    cls.HeartbeatSeconds = attr.ib(default=None, validator=optional(instance_of(int)))
-
-    return cls
+    Result = RHODES_ATTRIB()
 
 
 @attr.s(eq=False)
@@ -294,16 +183,18 @@ class Task(State):
     _required_fields = [RequiredValue("Resource", "Task resource is not set.")]
 
     # TODO: Additional validation for strings?
-    Resource = attr.ib(
-        default=None, validator=optional(instance_of((str, awslambda.Function, stepfunctions.Activity, GetAtt, Ref)))
-    )
+    Resource = RHODES_ATTRIB(validator=optional(instance_of(TASK_RESOURCE_TYPES)))
 
 
 @attr.s(eq=False)
+@_input_output
 class Choice(State):
     # TODO: Validate that Next and Default states are in parent
-    Choices: Iterable[ChoiceRule] = attr.ib(default=attr.Factory(list), validator=deep_iterable(member_validator=instance_of(ChoiceRule)))
-    Default: Optional[str] = attr.ib(default=None, validator=optional(instance_of(str)))
+    # TODO: Choice does not allow
+    Choices: List[ChoiceRule] = RHODES_ATTRIB(
+        default=attr.Factory(list), validator=deep_iterable(member_validator=instance_of(ChoiceRule))
+    )
+    Default: Optional[str] = RHODES_ATTRIB(validator=optional(instance_of(str)))
 
     def add_choice(self, rule: ChoiceRule) -> ChoiceRule:
         if rule.member_of is not None:
@@ -343,43 +234,83 @@ class Choice(State):
 
 
 @attr.s(eq=False)
+@_input_output
+@_next_and_end
 class Wait(State):
-    # TODO: Exactly one of these
-    Seconds: Optional[int] = attr.ib(default=None, validator=optional(instance_of(int)))
+    # TODO: Required only one of these on serialization
+
+    Seconds: Optional[int] = RHODES_ATTRIB(validator=optional(instance_of(int)))
     # TODO: Timestamp must be ISO8601 timestamp
     # TODO: Just only accept datetime objects, like Timestamp choice rules
-    Timestamp: Optional[str] = attr.ib(default=None)
-    SecondsPath: Optional[JsonPath] = attr.ib(default=None, validator=optional(instance_of(JsonPath)), converter=convert_to_json_path)
-    TimestampPath: Optional[JsonPath] = attr.ib(default=None, validator=optional(instance_of(JsonPath)), converter=convert_to_json_path)
+    Timestamp: Optional[str] = RHODES_ATTRIB()
+    SecondsPath: Optional[JsonPath] = RHODES_ATTRIB(
+        validator=optional(instance_of(JsonPath)), converter=convert_to_json_path
+    )
+    TimestampPath: Optional[JsonPath] = RHODES_ATTRIB(
+        validator=optional(instance_of(JsonPath)), converter=convert_to_json_path
+    )
+    _exactly_one = ValueError(
+        "Exactly one of 'Seconds', 'Timestamp', 'SecondsPath', and 'TimestampPath' must be supplied."
+    )
+
+    @Seconds.validator
+    def _check_timestamp(self, attribute: attr.Attribute, value: Any):
+        if value is None:
+            return
+
+        if not (self.Timestamp is self.SecondsPath is self.TimestampPath is None):
+            raise self._exactly_one
 
     @Timestamp.validator
-    def _check_timestamp(self, attribute, value):
+    def _check_timestamp(self, attribute: attr.Attribute, value: Any):
+        if value is None:
+            return
+
         if not is_valid_timestamp(value):
-            raise ValueError("Invalid Timestamp value: {}".format(value))
+            raise ValueError(f"Invalid Timestamp value: {value}")
+
+        if not (self.Seconds is self.SecondsPath is self.TimestampPath is None):
+            raise self._exactly_one
+
+    @SecondsPath.validator
+    def _check_timestamp(self, attribute: attr.Attribute, value: Any):
+        if value is None:
+            return
+
+        if not (self.Timestamp is self.Seconds is self.TimestampPath is None):
+            raise self._exactly_one
+
+    @TimestampPath.validator
+    def _check_timestamp(self, attribute: attr.Attribute, value: Any):
+        if value is None:
+            return
+
+        if not (self.Timestamp is self.Seconds is self.SecondsPath is None):
+            raise self._exactly_one
 
 
 @attr.s(eq=False)
 class Succeed(State):
     """"""
-    # TODO: Succeed does not accept any of the parent parameters...
 
 
 @attr.s(eq=False)
 class Fail(State):
     """"""
-    # TODO: The only parent parameter that Fail accepts is Comment
 
-    Error: Optional[str] = attr.ib(default=None, validator=optional(instance_of(str)))
-    Cause: Optional[str] = attr.ib(default=None, validator=optional(instance_of(str)))
+    Error: Optional[str] = RHODES_ATTRIB(validator=optional(instance_of(str)))
+    Cause: Optional[str] = RHODES_ATTRIB(validator=optional(instance_of(str)))
 
 
 @attr.s(eq=False)
 @_parameters  # TODO: Spec and developer guide disagree on whether this is accepted...
-@_result_path
 @_catch_retry
+@_result_path
+@_input_output
+@_next_and_end
 class Parallel(State):
     # TODO: Each branch MUST be a self-contained state machine.
-    Branches: Iterable[StateMachine] = attr.ib(default=attr.Factory(list))
+    Branches: List[StateMachine] = RHODES_ATTRIB(default=attr.Factory(list))
 
     def to_dict(self) -> Dict:
         self_dict = super(Parallel, self).to_dict()
@@ -389,7 +320,7 @@ class Parallel(State):
 
         return self_dict
 
-    def add_branch(self, state_machine: Optional[StateMachine] = None):
+    def add_branch(self, state_machine: Optional[StateMachine] = None) -> StateMachine:
         if state_machine is None:
             state_machine = StateMachine()
 
@@ -399,17 +330,19 @@ class Parallel(State):
 
 @attr.s(eq=False)
 @_parameters
-@_result_path
 @_catch_retry
+@_result_path
+@_input_output
+@_next_and_end
 class Map(State):
     _required_fields = [
         RequiredValue("Iterator", "Map iterator must be set."),
-        RequiredValue("ItemsPath", "Map items path must be set.")
+        RequiredValue("ItemsPath", "Map items path must be set."),
     ]
 
     # TODO: Iterator MUST be a self-contained state machine.
-    Iterator: StateMachine = attr.ib(default=None, validator=instance_of(StateMachine))
+    Iterator: StateMachine = RHODES_ATTRIB(validator=instance_of(StateMachine))
     # TODO: ItemsPath MUST be a valid JSON-path
-    ItemsPath: JsonPath = attr.ib(default=None, validator=optional(instance_of(JsonPath)), converter=convert_to_json_path)
+    ItemsPath: JsonPath = RHODES_ATTRIB(validator=optional(instance_of(JsonPath)), converter=convert_to_json_path)
     # TODO: MaxConcurrency MUST be non-negative
-    MaxConcurrency: Optional[int] = attr.ib(default=None, validator=optional(instance_of(int)))
+    MaxConcurrency: Optional[int] = RHODES_ATTRIB(validator=optional(instance_of(int)))