+++ src/rhodes/states/services/ecs.py
@@ -4,12 +4,14 @@ https://docs.aws.amazon.com/step-functions/latest/dg/connect-ecs.html
 """
 import attr
 
-from rhodes._util import RequiredValue
-from rhodes.states.services import IntegrationPattern, ServiceArn, ServiceIntegration, _supports_patterns
+from rhodes._util import RHODES_ATTRIB, RequiredValue
+from rhodes.states.services import ServiceIntegration
+from rhodes.states.services.identifiers import IntegrationPattern, ServiceArn
+from rhodes.states.services.util import supports_patterns
 
 
 @attr.s(eq=False)
-@_supports_patterns(
+@supports_patterns(
     IntegrationPattern.REQUEST_RESPONSE, IntegrationPattern.SYNCHRONOUS, IntegrationPattern.WAIT_FOR_CALLBACK
 )
 class AmazonEcs(ServiceIntegration):
@@ -19,12 +21,12 @@ class AmazonEcs(ServiceIntegration):
     # TODO: Sort out validation rules
     #  https://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_RunTask.html#ECS-RunTask-request-cluster
 
-    Cluster = attr.ib(default=None)
-    Group = attr.ib(default=None)
-    LaunchType = attr.ib(default=None)
-    NetworkConfiguration = attr.ib(default=None)
-    Overrides = attr.ib(default=None)
-    PlacementConstraints = attr.ib(default=None)
-    PlacementStrategy = attr.ib(default=None)
-    PlatformVersion = attr.ib(default=None)
-    TaskDefinition = attr.ib(default=None)
+    Cluster = RHODES_ATTRIB()
+    Group = RHODES_ATTRIB()
+    LaunchType = RHODES_ATTRIB()
+    NetworkConfiguration = RHODES_ATTRIB()
+    Overrides = RHODES_ATTRIB()
+    PlacementConstraints = RHODES_ATTRIB()
+    PlacementStrategy = RHODES_ATTRIB()
+    PlatformVersion = RHODES_ATTRIB()
+    TaskDefinition = RHODES_ATTRIB()