+++ src/rhodes/states/services/sqs.py
@@ -4,12 +4,14 @@ https://docs.aws.amazon.com/step-functions/latest/dg/connect-sqs.html
 """
 import attr
 
-from rhodes._util import RequiredValue
-from rhodes.states.services import IntegrationPattern, ServiceArn, ServiceIntegration, _supports_patterns
+from rhodes._util import RHODES_ATTRIB, RequiredValue
+from rhodes.identifiers import IntegrationPattern, ServiceArn
+from rhodes.states.services import ServiceIntegration
+from rhodes.states.services.util import supports_patterns
 
 
 @attr.s(eq=False)
-@_supports_patterns(IntegrationPattern.REQUEST_RESPONSE, IntegrationPattern.WAIT_FOR_CALLBACK)
+@supports_patterns(IntegrationPattern.REQUEST_RESPONSE, IntegrationPattern.WAIT_FOR_CALLBACK)
 class AmazonSqs(ServiceIntegration):
     _required_fields = (
         RequiredValue("MessageBody", "Amazon SQS Task requires a message body"),
@@ -20,9 +22,9 @@ class AmazonSqs(ServiceIntegration):
     # TODO: Sort out validation rules
     #  https://docs.aws.amazon.com/step-functions/latest/dg/connect-sqs.html
 
-    DelaySeconds = attr.ib(default=None)
-    MessageAttribute = attr.ib(default=None)
-    MessageBody = attr.ib(default=None)
-    MessageDeduplicationId = attr.ib(default=None)
-    MessageGroupId = attr.ib(default=None)
-    QueueUrl = attr.ib(default=None)
+    DelaySeconds = RHODES_ATTRIB()
+    MessageAttribute = RHODES_ATTRIB()
+    MessageBody = RHODES_ATTRIB()
+    MessageDeduplicationId = RHODES_ATTRIB()
+    MessageGroupId = RHODES_ATTRIB()
+    QueueUrl = RHODES_ATTRIB()